refactor(direct-control): extract notification types
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves only notification view/decision/queue public type ownership into packages/civ7-direct-control/src/play/notifications/view.ts and notification dismissal input/summary/result type ownership into packages/civ7-direct-control/src/play/notifications/dismissal-request.ts. Keeps facade type re-exports in index.ts and rewires package-internal notification type imports to concrete owners.

Preserves public type contracts and does not change notification materialization source strings, decision hint classification, maxNotifications behavior, dismissal command source, approval-first dismissal behavior, dismissal verification polling, CLI queue or bulk-dismiss policy, runtime proof status, telemetry, AI ingestion, hotseat runtime proof, CLI semantic projection, Effect/oRPC procedure-core work, or Task 2.9.4 matrix-row acceptance. Diplomacy/narrative/progression closeout, tactical, operation, ready, public procedure schema, and telemetry type ownership remain pending separate owner slices.

Validation:
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test -- play-notification-view.test.ts notification-dismissal.test.ts diplomacy-response.test.ts narrative-choice.test.ts public-api.test.ts autoplay-and-turn.test.ts
- bun run --cwd packages/cli test -- game/play/notification/hud.test.ts game/play/notification/queue.test.ts game/play/notification/dismiss.test.ts game/play/notification/dismiss-queue.test.ts game/play/priorities.test.ts game.play.rehydrate.test.ts game.play.diplomacy-response.test.ts game.play.narrative.test.ts game.play.end-turn.test.ts
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- ownership/no-cycle rg scans

Local package/type relocation proof only; no runtime/live-game, notification behavior, dismissal behavior, CLI queue policy, hotseat product-path, AI-ingestion, semantic CLI, telemetry, or procedure-core proof claimed.
