+++ openspec/changes/civ7-support-direct-control-modularization/tasks.md
@@ -433,6 +433,14 @@ runtime/direct-control claims.
         procedure schemas, operation/proof telemetry, hotseat runtime proof, AI
         ingestion, CLI semantic projection, Effect/oRPC procedure-core work,
         and Task 2.9.4 matrix-row acceptance pending.
+  - [x] 4.8.19 Extract notification view/dismissal type owners while keeping
+        public facade type re-exports in `index.ts`, preserving notification
+        view, decision hint/queue, dismissal input/summary/result public
+        contracts while leaving diplomacy/narrative/progression closeout,
+        tactical, operation, and ready result types, procedure schemas,
+        operation/proof telemetry, hotseat runtime proof, AI ingestion, CLI
+        semantic projection, Effect/oRPC procedure-core work, and Task 2.9.4
+        matrix-row acceptance pending.
 - [ ] 4.9 Import or explicitly cite oRPC architecture authority before any
       oRPC implementation; current support branch has no tracked
       `.agents/skills/civ7-orpc-control-architecture` or
