+++ openspec/changes/civ7-support-direct-control-modularization/workstream/direct-control-atom-corpus.md
@@ -215,6 +215,20 @@ Effect/oRPC procedure-core work, or Task 2.9.4 matrix-row acceptance. Tactical,
 operation, ready, public procedure schema, and telemetry type ownership remain
 pending separate owner slices.
 
+Notification type update: `src/play/notifications/view.ts` now owns play
+decision hint/input/action, notification summary, decision queue item, and
+notification view result types, and `src/play/notifications/dismissal-request.ts`
+now owns notification dismissal input/summary/result types while public facade
+type re-exports stay in `index.ts`. This is local package/type relocation proof
+only and does not change notification materialization source strings, decision
+hint classification, `maxNotifications` behavior, dismissal command source,
+approval-first dismissal behavior, dismissal verification polling, CLI queue or
+bulk-dismiss policy, runtime proof status, telemetry, AI ingestion, hotseat
+runtime proof, CLI semantic projection, Effect/oRPC procedure-core work, or Task
+2.9.4 matrix-row acceptance. Diplomacy/narrative/progression closeout, tactical,
+operation, ready, public procedure schema, and telemetry type ownership remain
+pending separate owner slices.
+
 Bounded root inspection update: `src/runtime/root-inspection.ts` now owns
 `inspectCiv7Root` and the generated bounded root inspection command while the
 public facade export stays in `index.ts` and injects command execution,
